moved ReadSource to own file, added some docstrings, misc useless PEP8 fixes
